+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Adds a one-time execution option (`once`) to `EventListener` [#32](https://github.com/TENSIILE/saborter/pull/32)
 - Adds an extended stack for the `AbortError` error [#33](https://github.com/TENSIILE/saborter/pull/33)
 - Adds a `reason` option to `TimeoutError` [#34](https://github.com/TENSIILE/saborter/pull/34)
+- Adds the `isAborted` flag to `Aborter` [#37](https://github.com/TENSIILE/saborter/pull/37)
 
 ### Bug Fixes